backend/rag-content-gen/requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
numpy==1.24.3
httpx==0.26.0